requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
black==24.2.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
coverage==7.4.3
crashtest==0.4.1
cryptography==42.0.5
distlib==0.3.8
dulwich==0.21.7
Faker==23.2.1
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.13.3
freezegun==1.4.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
iniconfig==2.0.0
installer==0.7.0
isort==5.13.2
jaraco.classes==3.3.1
jeepney==0.8.0
Jinja2==3.1.3
keyring==24.3.1
Mako==1.3.2
MarkupSafe==2.1.5
more-itertools==10.2.0
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.4.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
psycopg==3.1.18
psycopg-binary==3.1.18
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyproject_hooks==1.0.0
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
rapidfuzz==3.7.0
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
tomlkit==0.12.4
trove-classifiers==2024.3.25
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
zipp==3.18.1